Fix build with gssapi
krb5-dev has a hard requirement on e2fsprogs-dev, the provider of
libcom_err, so add the buildreq manually so that appropriate sasl gssapi
libs are built/installed.

Probably a good long-term solution here is for upstream krb5 to declare
"Requires: com_err" in the appropriate pkgconfig file(s), and then
pkgconfig() dependencies will be autogenerated for krb5-dev. I will make
note to investigate the feasibility of this solution.

Should resolve the latest errors from:
https://github.com/clearlinux/distribution/issues/945

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>